+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Function `Resolve-CIdentity`. Migrated to the `Carbon.Accounts` module.
 * Function `ConvertTo-CSecurityIdentifier`. Migrated to the `Carbon.Accounts` module.
+* Function `Get-CPathProvider`. Migrated to `Carbon.Core` module.
 
 
 ## 2.15.1